+++ src/store/index.js
@@ -1,5 +1,7 @@
 import Vue from 'vue'
 import Vuex from 'vuex'
+import { vuexfireMutations, firestoreAction } from 'vuexfire'
+import { firestore } from '@/plugins/firestore'
 
 Vue.use(Vuex)
 
@@ -8,7 +10,11 @@ export default new Vuex.Store({
     user: null,
     locale: null,
     pin: false,
-    online: null
+    online: null,
+    laundries: [],
+    flatmates: [],
+    categories: [],
+    temperatures: []
   },
   mutations: {
     login (state, user) {
@@ -22,9 +28,43 @@ export default new Vuex.Store({
     },
     setOnline (state, bool) {
       state.online = bool
-    }
+    },
+    ...vuexfireMutations,
   },
   actions: {
+    /**
+     * bind firestore data to state
+     */
+    bindFirestore: firestoreAction(({ bindFirestoreRef }) => {
+      bindFirestoreRef('laundries', firestore.collection('laundries'))
+      bindFirestoreRef('flatmates', firestore.collection('flatmates'))
+      bindFirestoreRef('categories', firestore.collection('categories'))
+      bindFirestoreRef('temperatures', firestore.collection('temperatures'))
+    }),
+    /**
+     * define transactions on firestore
+     */
+    addLaundry: firestoreAction((context, laundry) => {
+      return firestore.collection('laundries').add(laundry)
+    }),
+    addUser: firestoreAction ((context, user) => {
+      return firestore.collection('flatmates').add({ name: user })
+    }),
+    deleteLaundry: firestoreAction ((context, laundry) => {
+      return firestore.collection('laundries').doc(laundry.id).delete()
+    }),
+    updateLoaded: firestoreAction ((context, laundry) => {
+      return firestore.collection('laundries').doc(laundry.id).update({ loaded: laundry.loaded })
+    }),
+    updateParticipants: firestoreAction ((context, laundry) => {
+      return firestore.collection('laundries').doc(laundry.id).update({ participants: laundry.participants })
+    }),
+    startLaundry: firestoreAction ((context, laundry) => {
+      return firestore.collection('laundries').doc(laundry.id).update({ started: true })
+    }),
+    fillLaundry: firestoreAction ((context, laundry) => {
+      return firestore.collection('laundries').doc(laundry.id).update({ full: true })
+    })
   },
   modules: {
   }